# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API REST - Municipalidad Valle del Sol
  version: 1.0.0
  description: |
    Especificacion OpenAPI de la plataforma de gestion de emergencias.

    Incluye la API publica expuesta por el BFF y los endpoints internos de los
    microservicios de usuarios, reportes y monitoreo.
servers:
  - url: http://localhost:8080
    description: BFF local
  - url: http://localhost:8083
    description: ms-usuarios local
  - url: http://localhost:8081
    description: ms-reportes local
  - url: http://localhost:8082
    description: ms-monitoreo local
...
security:
  - bearerAuth: []
paths:
...
  /bff/reportes/{id}:
    get:
      tags: [BFF - Reportes]
      summary: Obtener reporte por ID
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Reporte encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reporte"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [BFF - Reportes]
      summary: Eliminar reporte
      description: Requiere token ADMIN.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Reporte eliminado.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /bff/reportes/{id}/estado:
    patch:
      tags: [BFF - Reportes]
      summary: Actualizar estado de reporte
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstadoUpdateRequest"
            example:
              estado: ATENDIDO
      responses:
        "200":
          description: Reporte actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reporte"
        "404":
          $ref: "#/components/responses/NotFound"
  /bff/reportes/media:
    post:
      tags: [BFF - Reportes]
      summary: Subir archivo multimedia para reporte
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo:
                  type: string
                  format: binary
      responses:
        "201":
          description: Archivo guardado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mediaUrl:
                    type: string
              example:
                mediaUrl: /bff/reportes/media/2bdc55b1-foto.jpg
        "400":
          $ref: "#/components/responses/BadRequest"
  /bff/reportes/media/{nombre}:
    get:
      tags: [BFF - Reportes]
      summary: Obtener archivo multimedia
      security: []
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
          example: 2bdc55b1-foto.jpg
      responses:
        "200":
          description: Archivo multimedia.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/notificaciones:
    get:
      tags: [ms-usuarios]
      summary: Listar notificaciones pendientes
      responses:
        "200":
          description: Notificaciones pendientes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notificacion"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/notificaciones/{id}/leida:
    patch:
      tags: [ms-usuarios]
      summary: Marcar notificacion como leida
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Notificacion marcada como leida.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/notificaciones/broadcast:
    post:
      tags: [ms-usuarios]
      summary: Crear notificacion para todos los usuarios
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastRequest"
            example:
              titulo: Nuevo reporte de emergencia
              mensaje: Se registro un reporte FORESTAL con intensidad ALTA.
              url: /
      responses:
        "201":
          description: Notificaciones creadas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  creadas:
                    type: integer
              example:
                creadas: 3
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/reportes:
    post:
      tags: [ms-reportes]
      summary: Crear reporte directamente en ms-reportes
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteCreateRequest"
      responses:
        "201":
          description: Reporte creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reporte"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [ms-reportes]
      summary: Listar reportes directamente en ms-reportes
      security: []
      responses:
        "200":
          description: Lista de reportes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reporte"
  /api/reportes/{id}:
    get:
      tags: [ms-reportes]
      summary: Obtener reporte por ID directamente en ms-reportes
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Reporte encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reporte"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [ms-reportes]
      summary: Eliminar reporte directamente en ms-reportes
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Reporte eliminado.
  /api/reportes/tipo/{tipo}:
    get:
      tags: [ms-reportes]
      summary: Filtrar reportes por tipo
      security: []
      parameters:
        - name: tipo
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/TipoReporte"
          example: FORESTAL
      responses:
        "200":
          description: Reportes filtrados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reporte"
  /api/reportes/estado/{estado}:
    get:
      tags: [ms-reportes]
      summary: Filtrar reportes por estado
      security: []
      parameters:
        - name: estado
          in: path
          required: true
          schema:
            type: string
          example: ACTIVO
      responses:
        "200":
          description: Reportes filtrados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reporte"
  /api/reportes/{id}/estado:
    patch:
      tags: [ms-reportes]
      summary: Actualizar estado de reporte directamente en ms-reportes
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstadoUpdateRequest"
            example:
              estado: ATENDIDO
      responses:
        "200":
          description: Reporte actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reporte"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/reportes/media:
    post:
      tags: [ms-reportes]
      summary: Subir archivo directamente a ms-reportes
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archivo]
              properties:
                archivo:
                  type: string
                  format: binary
      responses:
        "201":
          description: Archivo guardado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mediaUrl:
                    type: string
              example:
                mediaUrl: /bff/reportes/media/2bdc55b1-foto.jpg
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/reportes/media/{nombre}:
    get:
      tags: [ms-reportes]
      summary: Obtener archivo directamente desde ms-reportes
      security: []
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
          example: 2bdc55b1-foto.jpg
      responses:
        "200":
          description: Archivo multimedia.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  responses:
    BadRequest:
      description: Solicitud invalida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: La solicitud contiene datos invalidos.
    Unauthorized:
      description: Token ausente, invalido o credenciales incorrectas.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Debe enviar un token Bearer valido.
    Forbidden:
      description: Usuario sin permisos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Solo un administrador puede realizar esta accion.
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            message: Recurso no encontrado.
...
  schemas:
...
    Notificacion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titulo:
          type: string
        mensaje:
          type: string
        url:
          type: string
        creadaEn:
          type: string
          format: date-time
    BroadcastRequest:
      type: object
      required: [titulo, mensaje]
      properties:
        titulo:
          type: string
        mensaje:
          type: string
        url:
          type: string
          default: /
    TipoReporte:
      type: string
      enum: [FORESTAL, URBANO, SIMULACRO]
    Intensidad:
      type: string
      enum: [BAJA, MEDIA, ALTA, CRITICA]
    ReporteCreateRequest:
      type: object
      required: [tipo, descripcion, latitud, longitud]
      properties:
        tipo:
          $ref: "#/components/schemas/TipoReporte"
        descripcion:
          type: string
        latitud:
          type: number
          format: double
        longitud:
          type: number
          format: double
        mediaUrl:
          type: string
        intensidad:
          $ref: "#/components/schemas/Intensidad"
        sector:
          type: string
          description: Usado por el BFF para crear el foco automatico.
        ciudadanoId:
          type: string
        ciudadanoNombre:
          type: string
    Reporte:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tipo:
          $ref: "#/components/schemas/TipoReporte"
        descripcion:
          type: string
        latitud:
          type: number
          format: double
        longitud:
          type: number
          format: double
        mediaUrl:
          type: string
        intensidad:
          $ref: "#/components/schemas/Intensidad"
        estado:
          type: string
          example: ACTIVO
        fechaCreacion:
          type: string
          format: date-time
        ciudadanoId:
          type: string
        ciudadanoNombre:
          type: string
    EstadoUpdateRequest:
      type: object
      required: [estado]
      properties:
        estado:
          type: string
          example: ATENDIDO
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string